chore: remove obsolete workflow files (#21)
* chore: remove obsolete workflow files

deleted multiple workflow files for problems 1-7 and added new workflows for auto formatting and changed problems check.

* chore: update permissions for workflow

added permissions for contents and pull-requests in Auto Format workflow.

* chore: Auto format code

* chore: removed unnecessary usings

mainly to test the new workflows

* chore: fix commit message casing

corrected the casing of the commit message in the auto format workflow.

* chore: improve formatting detection logic

Refactor the commit step to check for changes more efficiently.

* chore: update file patterns for code formatting

changed file patterns in the auto format workflow to include '*.slnx'.