feat: added the problem 1. two sum (#23)
* fix: changed the attribute into valid

* fix: changed the variable into protected

* feat: add initial project and solution files

created TwoSum project with necessary configurations and added to solution.

* feat: add brute force and hash table solutions

implemented two approaches for the two sum problem with corresponding tests.

* feat: update README and add solutions overview

- updated C# solutions count in README
- created SOLUTIONS.md for problem categories and implemented solutions

* fix(devcontainer): update runArgs name for consistency

Changed the container name in runArgs to match the project.

* fix(workflow): moved and renamed C# auto formatter and solutions test

* fix(workflow): specify source directory for dotnet format

Updated the C# auto formatter workflow to explicitly format code in the 'src/csharp' directory.

* fix(sln): correct folder path format in solution file

Updated the folder name in the solution file to include a leading slash for proper path resolution.